create stop function
create reset benchbotData function, remove redundant code in traversal function
